requirements.txt
mistralai>=1.0.0
python-dotenv>=1.0.1
Pillow>=10.4.0
moviepy>=2.0.0
edge-tts>=6.1.12
pydantic>=2.8.2
